@@ -2,15 +2,15 @@
   "name": "portfolio-backend",
   "version": "1.0.0",
   "description": "Backend API for MD Mahbub Alam's portfolio website",
-  "main": "server.js",
+  "main": "server.ts",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
+    "start": "ts-node server.ts",
+    "dev": "nodemon --exec ts-node server.ts",
     "test": "jest",
     "test:watch": "jest --watchAll",
     "lint": "eslint .",
-    "seed": "node utils/seeder.js",
-    "seed:admin": "node utils/createAdminUser.js"
+    "seed": "ts-node utils/seeder.ts",
+    "seed:admin": "ts-node utils/createAdminUser.ts"
   },
   "keywords": [
     "portfolio",
@@ -23,30 +23,45 @@
   "license": "MIT",
   "dependencies": {
     "axios": "^1.8.2",
-    "bcryptjs": "^2.4.3",
+    "bcryptjs": "^3.0.2",
     "compression": "^1.8.0",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
-    "express-rate-limit": "^5.5.0",
-    "helmet": "^4.6.0",
+    "express-rate-limit": "^7.5.0",
+    "helmet": "^8.0.0",
     "jsonwebtoken": "^9.0.2",
-    "marked": "^4.3.0",
+    "marked": "^15.0.7",
     "memory-cache": "^0.2.0",
     "mongoose": "^8.12.1",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
-    "nodemailer": "^6.7.0",
+    "nodemailer": "^6.10.0",
     "rate-limiter-flexible": "^5.0.5",
-    "shortid": "^2.2.16",
+    "shortid": "^2.2.17",
     "slugify": "^1.6.6",
-    "uuid": "^8.3.2",
+    "uuid": "^11.1.0",
     "validator": "^13.12.0"
   },
   "devDependencies": {
-    "eslint": "^8.1.0",
-    "jest": "^27.5.1",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^5.0.0",
+    "@types/helmet": "^0.0.48",
+    "@types/jest": "^29.5.14",
+    "@types/jsonwebtoken": "^9.0.9",
+    "@types/mongoose": "^5.11.96",
+    "@types/morgan": "^1.9.9",
+    "@types/multer": "^1.4.12",
+    "@types/node": "^22.13.10",
+    "@types/supertest": "^6.0.2",
+    "@types/validator": "^13.12.2",
+    "eslint": "^9.22.0",
+    "jest": "^29.7.0",
     "nodemon": "^3.1.9",
-    "supertest": "^6.3.4"
+    "supertest": "^7.0.0",
+    "ts-jest": "^29.2.6",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.2"
   }
 }
